+++ client/src/queries/chat/getChat.queries.ts
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client';
+
+const GET_CHAT = gql`
+  query GetChat($chatId: String!) {
+    getChat(chatId: $chatId) {
+      result
+      chats {
+        writer
+        content
+        createdAt
+      }
+    }
+  }
+`;
+
+export default GET_CHAT;
