+++ client/src/queries/chat/index.ts
@@ -1,0 +1,3 @@
+export { default as CREATE_CHAT } from './createChat.queries';
+export { default as GET_CHAT } from './getChat.queries';
+export { default as SUB_CHAT } from './subChat.queries';
