+++ client/src/queries/chat/subChat.queries.ts
@@ -1,0 +1,17 @@
+import { gql } from '@apollo/client';
+
+const SUB_CHAT = gql`
+  subscription SubChat($chatId: String!) {
+    subChat(orderId: $chatId) {
+      result
+      error
+      chat {
+        createdAt
+        writer
+        content
+      }
+    }
+  }
+`;
+
+export default SUB_CHAT;
